# stormborne-booking-backend/docs/swagger.yaml
basePath: /
definitions:
...
  error.AppError:
    properties:
      code:
        type: string
      error: {}
      message:
        type: string
    type: object
...
  model.Movie:
    properties:
      duration:
        type: string
      genre:
        type: string
      id:
        type: string
      imdbRating:
        type: string
      imdbVotes:
        type: string
      name:
        type: string
      plot:
        type: string
      poster:
        type: string
      rated:
        type: string
    type: object
  model.Slot:
    properties:
      endTime:
        type: string
      id:
        type: integer
      name:
        type: string
      startTime:
        type: string
    type: object
...
  response.ShowResponse:
    properties:
      cost:
        type: number
      date:
        type: string
      id:
        type: integer
      movie:
        $ref: '#/definitions/model.Movie'
      slot:
        $ref: '#/definitions/model.Slot'
    type: object
  response.ShowtimeResponse:
    properties:
      availableSeats:
        type: integer
      cost:
        type: number
      endTime:
        type: string
      screenId:
        type: integer
      screenName:
        type: string
      showId:
        type: integer
      slotId:
        type: integer
      slotName:
        type: string
      startTime:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a skyfox
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Booking API
  version: "1.0"
paths:
...
  /movies/{movieId}/available-dates:
    get:
      description: get all dates that have shows for a given movie
      parameters:
      - description: movie id
        in: path
        name: movieId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      summary: Movie Available Dates
      tags:
      - Shows
  /movies/{movieId}/showtimes:
    get:
      description: get all showtimes for a movie on a given date
      parameters:
      - description: movie id
        in: path
        name: movieId
        required: true
        type: string
      - description: date (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ShowtimeResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      summary: Movie Showtimes
      tags:
      - Shows
  /movies/scheduled:
    get:
      description: get all movies that have upcoming scheduled shows
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Movie'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      summary: Scheduled Movies
      tags:
      - Shows
...
  /shows:
    get:
      consumes:
      - application/json
      description: get shows by date
      parameters:
      - description: Enter basic auth
        in: header
        name: Authorization
        required: true
        type: string
      - description: to get shows
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ShowResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      security:
      - BasicAuth: []
      summary: Shows
      tags:
      - Shows
...
swagger: "2.0"